refactoring #4246561
