+++ plugin/Controller/AdminViewController.php
@@ -38,11 +38,10 @@ use onOffice\WPlugin\Gui\AdminPageEstateUnitList;
 use onOffice\WPlugin\Gui\AdminPageEstateUnitSettings;
 use onOffice\WPlugin\Gui\AdminPageFormList;
 use onOffice\WPlugin\Gui\AdminPageFormSettingsMain;
-use onOffice\WPlugin\Gui\AdminPageModules;
 use onOffice\WPlugin\Record\RecordManagerReadForm;
 use onOffice\WPlugin\Types\FieldsCollection;
 use onOffice\WPlugin\Utility\__String;
-use onOffice\WPlugin\Utility\WPPluginChecker;
+use onOffice\WPlugin\WP\WPPluginChecker;
 use onOffice\WPlugin\WP\ListTableBulkActionsHandler;
 use Parsedown;
 use HTMLPurifier_Config;
