+++ plugin/Installer/DatabaseChanges.php
@@ -31,8 +31,8 @@ use onOffice\WPlugin\DataView\DataDetailViewHandler;
 use onOffice\WPlugin\Template\TemplateCall;
 use onOffice\WPlugin\Types\ImageTypes;
 use onOffice\WPlugin\DataView\DataSimilarView;
-use onOffice\WPlugin\Utility\WPPluginChecker;
 use onOffice\WPlugin\WP\WPOptionWrapperBase;
+use onOffice\WPlugin\WP\WPPluginChecker;
 use wpdb;
 use function dbDelta;
 use function esc_sql;
